+++ Nodes/backtest/Indicators.py
@@ -1,0 +1,7 @@
+
+
+class Indicators(object):
+    """
+    计算回测结果的指标信息
+    """
+    pass
