+++ Nodes/backtest/bt_utils.py
@@ -5,6 +5,16 @@ from numbers import Number
 
 import numpy as np
 import pandas as pd
+import uuid
+
+
+def random_str(seed=1, **kwargs):
+    # uln = 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789'
+    # rs = ''.join(random.sample(uln, num))  # 生成一个 指定位数的随机字符串
+    rs = ''
+    a = uuid.uuid1()  # 根据 时间戳生成 uuid , 保证全球唯一
+    b = rs + a.hex  # 生成将随机字符串 与 uuid拼接
+    return b  # 返回随机字符串
 
 
 def try_(lazy_func, default=None, exception=Exception):
